refactor: like/dislike 처리
