+++ src/comment/comment.module.ts
@@ -7,10 +7,8 @@ import { User } from 'src/user/entities/user.entity';
 import { Post } from 'src/post/entities/post.entity';
 import { UserModule } from 'src/user/user.module';
 import { AuthModule } from 'src/auth/auth.module';
-import { LikeModule } from 'src/like/like.module';
-import { DislikeModule } from 'src/dislike/dislike.module';
-import { CommentLike } from 'src/like/entities/comment-like.entity';
-import { CommentDislike } from 'src/dislike/entities/comment-dislike.entity';
+import { CommentLike } from 'src/comment/entities/comment-like.entity';
+import { CommentDislike } from 'src/comment/entities/comment-dislike.entity';
 import { AlarmModule } from 'src/alarm/alarm.module';
 import { PointModule } from 'src/point/point.module';
 import { Point } from 'src/point/entities/point.entity';
@@ -29,8 +27,6 @@ import { PointLog } from 'src/point/entities/point-log.entity';
     ]),
     UserModule,
     AuthModule,
-    LikeModule,
-    DislikeModule,
     AlarmModule,
     PointModule,
   ],
