+++ src/comment/comment.service.ts
@@ -1,4 +1,5 @@
 import {
+  BadRequestException,
   ForbiddenException,
   Injectable,
   NotFoundException,
@@ -11,8 +12,8 @@ import { Comment } from './entities/comment.entity';
 import { IsNull, Repository } from 'typeorm';
 import { Post } from 'src/post/entities/post.entity';
 import { User } from 'src/user/entities/user.entity';
-import { CommentLike } from 'src/like/entities/comment-like.entity';
-import { CommentDislike } from 'src/dislike/entities/comment-dislike.entity';
+import { CommentLike } from 'src/comment/entities/comment-like.entity';
+import { CommentDislike } from 'src/comment/entities/comment-dislike.entity';
 import { Role } from 'src/user/types/user-role.type';
 import { COMMENT_MESSAGE } from 'src/constants/comment-message.constant';
 import { AlarmService } from 'src/alarm/alarm.service';
@@ -255,5 +256,161 @@ export class CommentService {
       id: commentId,
       content: '삭제된 댓글입니다.',
     });
+  }
+
+  /** 댓글 좋아요 조회 API **/
+  async getCommentLikes(commentId: number) {
+    // 1. 댓글이 존재하니?
+    const existingComment = await this.commentRepository.findOneBy({
+      id: commentId,
+    });
+    // 1-1. 존재하지 않으면 에러처리
+    if (!existingComment) {
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
+    }
+
+    // 2. 총 댓글 좋아요 카운트
+    const count = await this.commentLikeRepository.countBy({ commentId });
+    return count;
+  }
+
+  /** 로그인한 사람의 댓글 좋아요 조회 API **/
+  async getMyCommentLike(userId: number, commentId: number) {
+    // 1. 댓글이 존재하는지?
+    const existingComment = await this.commentRepository.findOneBy({
+      id: commentId,
+    });
+    // 1-1. 존재하지 않으면 에러처리
+    if (!existingComment) {
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
+    }
+
+    // 2. 해당 댓글에 좋아요를 눌렀는지?
+    const like = await this.commentLikeRepository.findOneBy({
+      userId,
+      commentId,
+    });
+
+    return like ? true : false;
+  }
+
+  /** 댓글 좋아요 클릭 API **/
+  async clickCommentLike(userId: number, commentId: number) {
+    // 1. 좋아요를 누를 댓글 정보
+    const existingComment = await this.commentRepository.findOneBy({
+      id: commentId,
+    });
+    // 1-1. 댓글이 존재하지 않으면 에러처리
+    if (!existingComment) {
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
+    }
+    // 1-2. 본인의 댓글에는 좋아요를 누를 수 없도록
+    if (existingComment.userId === userId) {
+      throw new BadRequestException(
+        COMMENT_MESSAGE.LIKE.CREATE.FAILURE.NO_SELF
+      );
+    }
+
+    // 2. 내가 좋아요를 누른 상태인지 아닌지 확인
+    const commentLike = await this.commentLikeRepository.findOneBy({
+      userId: userId,
+      commentId: commentId,
+    });
+    if (!commentLike) {
+      // 2-1A. 댓글 좋아요 명단에 내가 없다면 => 댓글 좋아요 등록
+      await this.commentLikeRepository.save({
+        userId,
+        commentId,
+      });
+      // 2-1B. 댓글 좋아요에 따른 포인트 지급
+      const isValidPoint = await this.pointService.validatePointLog(
+        userId,
+        PointType.COMMENT_LIKE
+      );
+      if (isValidPoint) {
+        this.pointService.savePointLog(userId, PointType.COMMENT_LIKE, true);
+      }
+    } else {
+      // 2-2A. 댓글 좋아요 명단에 내가 있다면 => 댓글 좋아요 취소
+      await this.commentLikeRepository.delete({
+        userId,
+        commentId,
+      });
+      // 2-2B. 댓글 좋아요 취소에 따른 포인트 차감
+      this.pointService.savePointLog(userId, PointType.COMMENT_LIKE, false);
+    }
+  }
+
+  /** 댓글 싫어요 조회 API **/
+  async getCommentDislikes(commentId: number) {
+    // 1. 댓글이 존재하니?
+    const existingComment = await this.commentRepository.findOneBy({
+      id: commentId,
+    });
+    // 1-1. 존재하지 않으면 에러처리
+    if (!existingComment) {
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
+    }
+
+    // 2. 총 댓글 싫어요 카운트
+    const count = await this.commentDislikeRepository.countBy({ commentId });
+    return count;
+  }
+
+  /** 로그인한 사람의 댓글 싫어요 조회 API **/
+  async getMyCommentDislike(userId: number, commentId: number) {
+    // 1. 댓글이 존재하는지?
+    const existingComment = await this.commentRepository.findOneBy({
+      id: commentId,
+    });
+    // 1-1. 존재하지 않으면 에러처리
+    if (!existingComment) {
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
+    }
+
+    // 2. 해당 댓글에 싫어요를 눌렀는지?
+    const dislike = await this.commentDislikeRepository.findOneBy({
+      userId,
+      commentId,
+    });
+
+    return dislike ? true : false;
+  }
+
+  /** 댓글 싫어요 클릭 API **/
+  async clickCommentDislike(userId: number, commentId: number) {
+    // 1. 싫어요를 누를 댓글 정보
+    const existingComment = await this.commentRepository.findOneBy({
+      id: commentId,
+    });
+    // 1-1. 댓글이 존재하지 않으면 에러처리
+    if (!existingComment) {
+      throw new NotFoundException(COMMENT_MESSAGE.COMMENT.NOT_FOUND);
+    }
+    // 1-2. 본인의 댓글에는 싫어요를 누를 수 없도록
+    if (existingComment.userId == userId) {
+      throw new BadRequestException(
+        COMMENT_MESSAGE.DISLIKE.CREATE.FAILURE.NO_SELF
+      );
+    }
+
+    // 2. 내가 싫어요를 누른 상태인지 아닌지 확인
+    const commentDislike = await this.commentDislikeRepository.findOneBy({
+      userId,
+      commentId,
+    });
+    if (!commentDislike) {
+      // 2-1. 댓글 싫어요 명단에 내가 없다면 => 댓글 싫어요 등록
+      await this.commentDislikeRepository.save({
+        userId,
+        commentId,
+      });
+    } else {
+      // 2-2. 댓글 싫어요 명단에 내가 있다면 => 댓글 싫어요 취소
+      await this.commentDislikeRepository.delete({
+        userId,
+        commentId,
+      });
+    }
   }
 }
