+++ src/comment/entities/comment-dislike.entity.ts
