--- src/dislike/dtos/create-dislike.dto.ts
@@ -1,1 +1,0 @@
-export class CreateDislikeDto {}
