--- src/dislike/dtos/update-dislike.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { CreateDislikeDto } from './create-dislike.dto';
-
-export class UpdateDislikeDto extends PartialType(CreateDislikeDto) {}
