--- src/like/dtos/create-like.dto.ts
@@ -1,1 +1,0 @@
-export class CreateLikeDto {}
