--- src/like/dtos/update-like.dto.ts
@@ -1,4 +1,0 @@
-import { PartialType } from '@nestjs/swagger';
-import { CreateLikeDto } from './create-like.dto';
-
-export class UpdateLikeDto extends PartialType(CreateLikeDto) {}
