+++ src/post/post.module.ts
@@ -9,10 +9,20 @@ import { AwsModule } from 'src/aws/aws.module';
 import { PointModule } from 'src/point/point.module';
 import { Point } from 'src/point/entities/point.entity';
 import { PointLog } from 'src/point/entities/point-log.entity';
+import { PostLike } from './entities/post-like.entity';
+import { PostDislike } from './entities/post-dislike.entity';
 
 @Module({
   imports: [
-    TypeOrmModule.forFeature([Post, PostImage, User, Point, PointLog]),
+    TypeOrmModule.forFeature([
+      Post,
+      PostImage,
+      User,
+      Point,
+      PointLog,
+      PostLike,
+      PostDislike,
+    ]),
     AwsModule,
     PointModule,
   ],
