+++ src/user/entities/user.entity.ts
@@ -11,10 +11,10 @@ import {
 import { Role } from '../types/user-role.type';
 import { Comment } from 'src/comment/entities/comment.entity';
 import { Post } from 'src/post/entities/post.entity';
-import { CommentLike } from 'src/like/entities/comment-like.entity';
-import { CommentDislike } from 'src/dislike/entities/comment-dislike.entity';
-import { PostLike } from 'src/like/entities/post-like.entity';
-import { PostDislike } from 'src/dislike/entities/post-dislike.entity';
+import { CommentLike } from 'src/comment/entities/comment-like.entity';
+import { CommentDislike } from 'src/comment/entities/comment-dislike.entity';
+import { PostLike } from 'src/post/entities/post-like.entity';
+import { PostDislike } from 'src/post/entities/post-dislike.entity';
 import { Point } from 'src/point/entities/point.entity';
 import { PointLog } from '../../point/entities/point-log.entity';
 import { ChatRoom } from 'src/chat/entities/chat-room.entity';
